requirements.txt
et-xmlfile==1.1.0
lxml==4.6.3
openpyxl==3.0.7
Pillow==8.3.1
python-pptx==0.6.19
XlsxWriter==3.0.1
